ControllerTestの追加

+++ src/main/java/com/example/BankApp/dto/BankAccountResponse.java
@@ -13,5 +13,5 @@ public class BankAccountResponse {
 
   private String accountNumber;
   private String accountHolderName;
-  private double balance;
+  private int balance;
 }
