+++ src/main/java/com/example/BankApp/model/BankAccount.java
@@ -17,6 +17,6 @@ public class BankAccount {
   @Id
   private String accountNumber;
   private String accountHolderName;
-  private double balance;
+  private int balance;
   private boolean isActive;
 }
